generalize LandingButton to Button
