+++ src/components/Button/Button.jsx
@@ -1,10 +1,29 @@
-import React from "react";
+import React, { useState } from "react";
 import "./Button.scss";
+import pixelHeart from "../../resources/images/icons/pixel-heart.svg";
+import arrow from "../../resources/images/icons/arrow.svg";
 
-const Button = ({ label, link, type }) => (
-        <a className={type ?? "button"} href={link}>
-            {label}
-        </a>
-);
+const Button = ({ label, link, type }) => {
+
+  const [clicked, setClicked] = useState(false);
+  
+  const handleClick = () => {
+    setClicked(true);
+  };
+  
+  return (
+    <a
+    href={link}
+    className={`button ${type ?? ""} ${clicked ? "clicked" : ""}`}
+    target="_blank"
+    rel="noreferrer noopener"
+    onClick={handleClick}
+    >
+      <img className="pixel" src={pixelHeart} alt="Pixel Heart" />
+      <div className="label">{label}</div>
+      <img className="arrow" src={arrow} alt="Arrow" />
+    </a>
+  );
+};
 
 export default Button;
