+++ src/components/Button/Button.scss
@@ -1,34 +1,46 @@
 @import "../../constants";
 
-.button, .button-2 {
-    padding: 15px 30px;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    background-color: $accent-pink;
-    color: $text-light;
-    font-size: $text-small;
-    font-family: "Poppins";
-    font-weight: 700;
-    border-radius: 10px;
-    border-style: none;
-    text-decoration: none;
-    @include phone {
-        font-size: 16px;
-      }
-    
-    &:hover,
-    &:focus {
-        opacity: 0.7;
-        box-shadow: 0 0 11px rgba(255, 96, 120, 0.5);
-        transform: translateY(-5px);
-        transition: 0.1s linear;
-    }
-}
+.button {
+  padding: 13px;
+  display: inline-flex;
+  align-items: center;
+  height: auto;
+  gap: 10px;
+  background-color: white;
+  color: $text-dark;
 
-.button-2 {
-    background-color: transparent;
-    color: $accent-pink;
-    border: 2px solid $accent-pink;
-}
+  font-size: 16px;
+  font-family: "Poppins";
+  font-weight: 700;
+  text-decoration: none;
+  border-style: solid;
+  white-space: nowrap;
+  box-shadow: 10px 10px 0px 0px #19123c;
+  @include phone {
+    font-size: 14px;
+  }
 
+  &:hover,
+  &:focus {
+    opacity: 0.7;
+    box-shadow: 0 0 11px rgba(255, 96, 120, 0.5);
+    transform: translateY(-5px);
+    transition: 0.1s linear;
+  }
+  
+  &.clicked {
+    /* Remove hover and focus styles after clicking the button */
+    opacity: 1;
+    box-shadow: 10px 10px 0px 0px #19123c;
+    transform: none;
+    transition: none;
+  }
+
+  img {
+    width: 20px;
+  }
+
+  .label {
+    margin-right: auto;
+  }
+}
